requirements.txt
genesis-world
numpy
matplotlib
opencv-python
torch
torchvision
transformers

# 데이터 처리
h5py>=3.8.0
pandas>=1.5.0
scipy>=1.9.0
datasets>=2.14.0

# 설정 관리
pyyaml>=6.0

# 모니터링
psutil>=5.9.0

# 자연어 처리
spacy>=3.4.0
nltk>=3.8

# 실험 추적 (선택적)
mlflow>=2.0.0

# 웹 API (선택적)
fastapi>=0.95.0
uvicorn>=0.20.0

# 테스팅
pytest>=7.0.0
pytest-asyncio>=0.21.0